task: add TaskRole for tracking client/server role
The Task and its subclasses already track the full "ts" (TestSettings)
variable. Currently they also track a "node_name", which is always one
of "ts.conf_{server,server_used,client}.name".

Instead of passing on a redundant "node_name", the task should know its
role. If it knows it is for example a TaskRole.SERVER, it can
automatically determine the node name. Knowing its role is more useful
while also clearer.

It's clearer because the role is just one of a few TaskRole values, that
you can review whether they make sense. Based on that role, in the next
commit we will compute the "node_name". That is clearer, because you can
look at one place how the node name is computed based on the role.
Before, this was not easy to see, and probably also done wrongly as
TaskRole.SERVER is probably never the right one (rather
TaskRole.SERVER_USED).

Add a TaskRole and pass it to the Task object. For now, we still pass a
redundand node_name. That shall change next.

Signed-off-by: Thomas Haller <thaller@redhat.com>