task: handle "persistent" flag for tests running on the same node
For some tests (self.test_case_id.info.is_same_node), the server node is
actually the same as the client's. See TestSettings.node_server and
TestSettings._node_server.

This means, we almost always should use TestSettings.node_server and not
TestSettings._node_server.

Note that the test configuration contains a "persistent" flag for server
nodes. That flag does not exist for client nodes and for client nodes it
is always False (not implemented).

Use the right persistent flag for those tests that run on one node.

Signed-off-by: Thomas Haller <thaller@redhat.com>