tftbase: mark duplicate TestCaseType entries as aliases
A few TestCaseType exhibit the identical behavior as some others. We can
see that, because their TestCaseTypInfo contains the identical settings.
We also have a unit test that uncovers this ("identical_cases") and
which guards against adding such duplicate tests in the future.

Add a flag TestCaseTypInfo.deprecated_alias_for to those duplicated.
Later, we will treat them special.

Signed-off-by: Thomas Haller <thaller@redhat.com>