tft: skip test case types that are deprecated aliases
Some TestCaseType are identical to others. It makes no sense to run
them. However, we get them easily when specifying a range or "*".

If we have such a deprecated alias, then skip it if-and-only-if the
original value is also present.

Note that this skipping does not cover any further check for duplicates. With
ranges, we can easily specify the same test multiple times, and we will
run them. We only skip the aliases if the orignal is already selected.

Signed-off-by: Thomas Haller <thaller@redhat.com>