ktoolbox: bump to latest version
There isn't a relevant change there. This is only to sync
the used version with the latest available version.

What this however brings is that the ktoolbox version was bumped from
0.1.0 to 0.2.0. That is important, because with this a `pip install`
will actually upgrade the old version. Otherwise, if the version stays
at 0.1.0, a `pip install $URL` of the same version does not actually
upgrade the package.

Signed-off-by: Thomas Haller <thaller@redhat.com>