testConfig: validate kubeconfig files in two separate phases
TestConfig parses the configuration. Configuration can contain file
paths (like the kubeconfig), which may be invalid depending on what we
have on the file system.

We want that parsing the configuration does not yet validate the system
beyond the configuration itself. In a first step, the TestConfig()
constructor should only parse the configuration and input parameters,
but not yet look at the file system. The benefit is that we can use
TestConfig() on another system to only validate the configuration. For
that first step validation, we don't yet check for further files on the
system.

In a second step, we then often want to validate the input beyond that.
Add TestConfig.system_check() for that. This validates whether further
run time dependencies are present (like the kubeconfig files on disk).

Change the TestConfig() constructor to not yet fail if kubeconfig files
are invalid. Add (and call) a TestConfig.system_check(), which then
performs the second step of the validation.

TestConfig.system_check() could be further extended, for example to
check whether we have kubectl command in the path.

Why does TestConfig() even bother with this? True, we could not have
TestConfig.system_check() and let another component validate the
kubeconfig (and provide what TestConfig.client() does). But that just
moves the code somewhere else. As TestConfig already can contain the
kubeconfig paths in the YAML configuration, it makes sense that it also
is concerned with detecting the default values and providing the
TestConfig.client() instances. For that, we have those two steps both in
TestConfig. While this code could be somewhere else, it also makes sense
to have it here.

Note that the accessors like TestConfig's kubeconfig, kubeconfig_infra
or client() still only ever return one value. While that value gets
initialized lazily on first access (or during system_check()), once
initialized, it's cached and always the same value is returned. That
makes the accessors behave as if they are stateless (users only see a
single value whenever they call the accessors).

Another benefit is that the error messages for invalid kubeconfig files
is now better.

Signed-off-by: Thomas Haller <thaller@redhat.com>