tft: support manifests/overrides for providing alternative manifest files
Under "manifests" directory we have various files that we render. Add a
mechanism so that the user can provide alternative files.

The user can set TFT_MANIFEST_OVERRIDES environment variable, which can
be an overrides directory. If unset, it defaults to "manifests/overrides/".
If empty, no overrides are used.

If tft looks for manifest files to render (like "pod.yaml.j2"), it now
first looks at the overrides directory. If the file exists, it is taken.
It shadows the file that we commit to git. Otherwise, tft takes the file
from "manifests/" directory, as before.

Signed-off-by: Thomas Haller <thaller@redhat.com>