tft: ignore error from podman-rm for non-existing container
Newer version of podman exit with success for `podman rm --force`, if
the pod does not exist in the first place.

Older versions would fail with an error.

  # podman rm --force foo
  Error: no container with name or ID "foo" found: no such container
  # echo $?
  1

This leads to a WARNING message in the log.

This warning is not useful and distracts from actual problems. That is, because
we don't want to first whether the container exists before deleting it.
Suppress that failure.

Signed-off-by: Thomas Haller <thaller@redhat.com>