task: set "port" in parent implementation Task.get_template_args()
I think we should aim that the set of variables that we render is always
the same. That is, Task.get_template_args() renders most variables,
including "port".

Now, "port" should be set for ServerTask to the actual port number, and
for ClientTask to empty. This preserves previous behavior.

What we change here is that we also set the variable for "PluginTask"
subclasses (to empty). Note that PluginTask only renders
"tools-pod.yaml.j2", which does not use the variable so the change does
not affect them.

I find it clearer that we see all variables ("port") listed in
Task.get_template_args(). We can easily find to which value each task
implementation sets the variable (_get_template_args_port()).

Signed-off-by: Thomas Haller <thaller@redhat.com>