task: let parent implementation Task.get_template_args() set "pod_name"
The Task has a "pod_name" attribute. Arguably, Task.__init__() initializes
it only to "", and expects subclasses to set a suitable name. In practice,
all Task instances will have a valid self.pod_name.

Also, we render several templates, and while not every template may require
a {{pod_name}}, it is hard to only set the variable if-and-only-if the template
requires it.

Hence, always set it in the parent implementation. It works fine in any case,
even if the template ends up not accessing the {{pod_name}} variable.

Signed-off-by: Thomas Haller <thaller@redhat.com>