task: log invalid rendered manifest in Task.render_file()
After rendering the manifest file, we want to parse and print it
in debug logging. That is very useful to see what we did.

Jinja2 is plain text replacement and unaware that the template
is supposed to end up as valid YAML.

Catch errors while re-parsing the rendered file as YAML and log an error
message with the full content before failing.

Signed-off-by: Thomas Haller <thaller@redhat.com>