scripts: add "build-container.sh" script
This script is useful to build and push a container for testing. While
we automatically build and publish the container when pushing to git
main branch, it is useful to apply a patch for testing and build our own
containers.

You can use your own container from tft by setting TFT_TEST_IMAGE
environment variable.

Signed-off-by: Thomas Haller <thaller@redhat.com>