simple: add "--exec" option to run another script for testing
One point of the "simple" test is to allow to hack the test and do
whatever we want (for example, modify the python script).

However, the simple script that we run is still inside the container.
So to modify the script, you would have to rebuild the container and
select it via TFT_TEST_IMAGE environment variable. That is still
cumbersome.

Now, add an "--exec" option. If set, the python script will download
that URL, make it executable and execute it. This way, you can more
easily inject another script to run during the test.

You can specify both "--exec" and "--exec-args" option in tft via
"tft[0].connections[0].{server,client}.args" parameter. For example:

tft:
  - name: ...
    ...
    connections:
      - name: ...
        type: "simple"
        server:
          - name: ...
        client:
          - name: ...
            args: "--exec https://my/script.sh --exec-args hello"
Signed-off-by: Thomas Haller <thaller@redhat.com>