simple: accept empty "--exec" option
A reasonable thing to do for the exec script is to print some debugging
information about the call environment (`ip addr`) and then execute the
original script.

It can do so, by getting the original command line via the "ORIG_ARGS_*"
variables. However, those also contain the "--exec" option. If called
that way, the script will recursively call itself.

Instead, allow the script to overwrite the "--exec" option so it can do:

  "$ORIG_ARGS_0" "${OARGS[@]}" "--exec" ""

For that, accept an empty "--exec" option to reset a previously set
option.

Signed-off-by: Thomas Haller <thaller@redhat.com>