tft: log test result and duration at the end
While we can run `time ./tft.py` or even just scroll up and look at the
timestamps in the logfile, it is useful to see the duration at one
glance. Log that.

Also, we now always call print_results.process_results_all(). However,
depending on "--check" we might not log anything. The purpose is to
always collect the test result status. We might not use the result
status for the exit code (that depends on the "--check" option). But we
use the result status for the last message to show success or failure.

Signed-off-by: Thomas Haller <thaller@redhat.com>