task: support configuring test pods as priviledged
Usually we don't need this, and usually we want to run with restricted
permissions to be closer to what the user would do.

However, for testing it can be useful (e.g. to exec tcpdump in a running
container). Make it configurable via configuration YAML and
TFT_PRIVILEGED_POD environment variable.

Signed-off-by: Thomas Haller <thaller@redhat.com>