task: make "node_name_sanitized" a property
This seems more pythonic. The getter has no side effects and is cheap to
evaluate.

Also, it is based on the "node_name". That property is based on
immutable attributes and thus  yields the identical reference on each
call. This makes it essentially stateless and simple to reason about. As
such, "node_name_sanitized" also gives always the same value, albeit a
newly allocated string reference.