task: simplify selection of pod template and rename pod files
The if checks had duplicated code. For example, they all called
tftbase.get_manifest(), which can be done at the end of the block.

Also, we already choose a unique and suitable name for the pod. For the
generated YAML file, we should just use the same name. Previously, the
generated name would often differ from the pod name for no reason.