manifests: quote the {node,pod}_name Jinja variables in YAML
Jinja is not aware of what it is templating. So it does not know that we
want to set a string in YAML, or how to quote that.

Luckily, we already ensure that the node name is a valid DNS name and
the pod name is even more strict. So they cannot contain quotes that
could break our YAML. Still, a node name "1" would be interpreted as a
plain number in YAML. We always want a string.

Do the quoting.

In this case, this is safe, because we know that the node name and the
pod name only contain certain characters.